# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a water station.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "https://poseidon-backend.fly.dev/",
    "basePath": "/",
    "paths": {
        "/bottles": {
            "get": {
                "description": "Get all bottles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bottles"
                ],
                "summary": "Show all bottles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Bottle"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing bottle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bottles"
                ],
                "summary": "Update a bottle",
                "parameters": [
                    {
                        "description": "Bottle",
                        "name": "bottle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Bottle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Bottle"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new bottle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bottles"
                ],
                "summary": "Create a bottle",
                "parameters": [
                    {
                        "description": "Bottle",
                        "name": "bottle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Bottle"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.Bottle"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing bottle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bottles"
                ],
                "summary": "Delete a bottle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bottle ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/bottles/preferences/{nfc-id}": {
            "get": {
                "description": "Get bottle preferences by the NFC ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bottles"
                ],
                "summary": "Get bottle preferences by the NFC ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NFC ID",
                        "name": "nfc_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Bottle"
                        }
                    }
                }
            }
        },
        "/consumer_test_answers": {
            "get": {
                "description": "Get all consumer test answers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "consumer_test_answers"
                ],
                "summary": "Show all consumer test answers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.ConsumerTestAnswer"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing consumer test answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "consumer_test_answers"
                ],
                "summary": "Update a consumer test answer",
                "parameters": [
                    {
                        "description": "Consumer Test Answer",
                        "name": "answer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.ConsumerTestAnswer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.ConsumerTestAnswer"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new consumer test answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "consumer_test_answers"
                ],
                "summary": "Create a consumer test answer",
                "parameters": [
                    {
                        "description": "Consumer Test Answer",
                        "name": "answer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.ConsumerTestAnswer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.ConsumerTestAnswer"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing consumer test answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "consumer_test_answers"
                ],
                "summary": "Delete a consumer test answer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Consumer Test Answer ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/contribution/user": {
            "get": {
                "description": "Get the total water amount and savings for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contribution"
                ],
                "summary": "Get user contribution",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.WaterTransaction"
                        }
                    }
                }
            }
        },
...
        "/users/{userId}/bottles": {
            "get": {
                "description": "Get all bottles associated with a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bottles"
                ],
                "summary": "Get all bottles by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Bottle"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "database.Bottle": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "bottle_image": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "fill_volume": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "nfc_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "water_type": {
                    "type": "string"
                }
            }
        },
...
        "database.ConsumerTestAnswer": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "database.WaterTransaction": {
            "type": "object",
            "properties": {
                "bottle_id": {
                    "type": "integer"
                },
                "guest": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "station_id": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "volume": {
                    "type": "integer"
                },
                "water_type": {
                    "type": "string"
                }
            }
        }
    }
}